requirements.txt
geopandas
pandas
sqlalchemy
streamlit
matplotlib
pyyaml